// apps/api/src/types/qbittorrent.types.ts
export interface QBittorrentTorrent {
  hash: string;
  name: string;
  size: number;
  progress: number; // 0-1
  dlspeed: number; // bytes/s
  upspeed: number; // bytes/s
  priority: number;
  num_seeds: number;
  num_complete: number;
  num_leechs: number;
  num_incomplete: number;
  ratio: number;
  eta: number; // seconds
  state: string;
  seq_dl: boolean;
  f_l_piece_prio: boolean;
  category: string;
  tags: string;
  super_seeding: boolean;
  force_start: boolean;
  save_path: string;
  added_on: number; // unix timestamp
  completion_on: number;
  tracker: string;
  dl_limit: number;
  up_limit: number;
  downloaded: number;
  uploaded: number;
  downloaded_session: number;
  uploaded_session: number;
  amount_left: number;
  save_path_id?: number;
  time_active: number;
  seeding_time: number;
}

export interface QBittorrentServerState {
  connection_status: string;
  dht_nodes: number;
  dl_info_data: number;
  dl_info_speed: number;
  dl_rate_limit: number;
  up_info_data: number;
  up_info_speed: number;
  up_rate_limit: number;
  queueing: boolean;
  use_alt_speed_limits: boolean;
  refresh_interval: number;
}